reformatted to google style standard
This change was made for readability and code quality assurance.
