--- src/controller/ClientController.cpp
@@ -1,11 +1,0 @@
-#include "ClientController.h"
-#include "ViewFactory.h"
-
-using namespace chat_client_controller;
-
-ClientController::ClientController(int argc, char **argv) {
-     ViewFactory factory;
-     this->view = factory.createView(argc, argv);
-
-     this->view->waitForInternalThreadToExit();
-}
