+++ src/main.cc
@@ -1,0 +1,7 @@
+#include "controller/client-controller.h"
+
+using namespace chat_client_controller;
+
+int main(int argc, char **argv) {
+  ClientController clientController(argc, argv);
+}
