--- src/main.cpp
@@ -1,7 +1,0 @@
-#include "controller/ClientController.h"
-
-using namespace chat_client_controller;
-
-int main(int argc, char **argv) {
-    ClientController clientController(argc, argv);
-}
